requirements.txt
numpy~=1.26.2
rasterio~=1.3.9
Django~=5.0.1
djangorestframework~=3.14.0